Various improvements to setup_gen + added setup:keep_release(RelVsn)
The function setup:keep_release(RelVsn) takes a snapshot of the
currently loaded and running applications in the node, derives a
diff between the application environment and what is in the corresponding
.app files, and generates a setup config, .rel file and boot scripts
in a releases/RelVsn directory (assuming it has write permission to
do so).

The first purpose of this function was to allow you to 'save' a
set of app-level upgrades using setup:reload_app/[1,2] - otherwise,
the old versions of those applications would be started again after
a reboot. A few interactions with the release_handler would be needed
to make the generated release permanent - this is future work.

An idea that grew while working on this, was that it might be a good
way to interactively define a system, and then capture it for reuse.