Add post hook handling for rebar3
Add special handling in the rebar.config.script file for
rebar3. Currently rebar3 does not have escriptize support and a project
using rebar3 and an dependent application that relies on the setup
application fails to build when the post hook attempts to execute. If
the escript script file name has a base of rebar3, revert to the default
behavior specificed in the rebar.config file of using make as the
command.